Added tailwind CSS and Image component instead of vanilla CSS
