--- src/components/project-card/index.css
@@ -1,41 +1,0 @@
-.project-card {
-  align-items: center;
-  background-color: #3b7227b2;
-  border-radius: 20px;
-  border: solid 7px #529546;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  height: 100%;
-  justify-content: center;
-  padding: 20px 10px 20px 10px;
-  width: 100%;
-}
-
-.project-image-container {
-  align-items: flex-end;
-  background-position: center;
-  background-size: cover;
-  border-radius: 20px;
-  display: flex;
-  font-family: "Malayalam MN";
-  justify-content: flex-end;
-  margin-bottom: 10px;
-  min-height: 55%;
-  padding: 2px 20px;
-  text-align: right;
-  text-transform: uppercase;
-  width: 75%;
-}
-
-.description {
-  align-items: center;
-  color: white;
-  display: flex;
-  font-family: "Malayalam MN";
-  font-weight: 400;
-  line-height: 15px;
-  size: 10px;
-  text-align: center;
-  width: 80%;
-}
